Explicitely name segments private networks
The Vagrant private networks used to represent segments are explicitely
name to have better control of where the VMs that constitute this
environment are connected. The interfaces to those networks are also
named to facilitate debugging.

The routed network segments naming convention is updated to better
support development of the multiple segments per host functionality.